requirements.txt
Django
Fabric
apache-libcloud==0.9.1
